refactor: clean up viewer hooks
