--- packages/viewer/src/hooks.tsx
@@ -1,360 +1,0 @@
-import React from "react";
-import { ErrorBoundary } from "react-error-boundary";
-import type {
-  ReactCADCore,
-  ReactCADView,
-  ReactCADNode,
-  ProgressIndicator,
-} from "@react-cad/core";
-import { createRoot } from "@react-cad/renderer";
-import { ExportFns, ViewOptions } from "./types";
-
-import DetailContext from "./DetailContext";
-
-const NullComponent = () => null;
-
-const originalConsoleError = console["error"];
-
-const subscribers: any[] = [];
-
-console["error"] = (message) => {
-  if (message.startsWith("The above error")) {
-    subscribers.forEach((subscriber) => subscriber(message));
-  }
-  originalConsoleError(message);
-};
-
-export function useReactCADRenderer(
-  core: ReactCADCore,
-  shape: React.ReactElement,
-  detail: ViewOptions["detail"],
-  reset?: boolean
-): [ReactCADNode, number, boolean, Error | undefined, string | undefined] {
-  const rootNode = React.useMemo(() => core.createCADNode("union"), [core]);
-  const root = React.useMemo(() => createRoot(rootNode, core), [core]);
-
-  const [renderFrameId, setRenderFrameId] = React.useState(0);
-  const [renderReset, setRenderReset] = React.useState(false);
-  const [renderError, setRenderError] = React.useState<Error>();
-  const [renderErrorContext, setRenderErrorContext] = React.useState<string>();
-
-  const renderCallback = React.useMemo(() => {
-    let shouldReset = Boolean(reset);
-
-    return async () => {
-      setRenderReset(shouldReset);
-      setRenderFrameId((n) => n + 1);
-
-      shouldReset = false;
-    };
-  }, [core, rootNode, reset]);
-
-  React.useEffect(() => {
-    setRenderError(undefined);
-    setRenderErrorContext(undefined);
-
-    root.render(
-      <ErrorBoundary
-        FallbackComponent={NullComponent}
-        onError={setRenderError}
-        resetKeys={[shape]}
-      >
-        <DetailContext.Provider value={detail}>{shape}</DetailContext.Provider>
-      </ErrorBoundary>,
-      renderCallback
-    );
-  }, [shape, detail, root, rootNode]);
-
-  React.useEffect(() => {
-    subscribers.push(setRenderErrorContext);
-    return () => {
-      const index = subscribers.findIndex(setRenderErrorContext);
-      if (index >= 0) {
-        subscribers.splice(index, 1);
-      }
-      root.delete();
-      rootNode.delete();
-    };
-  }, []);
-
-  return [
-    rootNode,
-    renderFrameId,
-    renderReset,
-    renderError,
-    renderErrorContext,
-  ];
-}
-
-export function useReactCADView(
-  canvasContainerRef: React.RefObject<HTMLDivElement>,
-  core: ReactCADCore,
-  options: ViewOptions,
-  addTask: AddTask
-): React.MutableRefObject<ReactCADView | undefined> {
-  const canvasRef = React.useRef<HTMLCanvasElement>();
-  const view = React.useRef<ReactCADView>();
-  const [loaded, setLoaded] = React.useState(false);
-
-  React.useEffect(() => {
-    canvasRef.current = canvasContainerRef.current?.getElementsByTagName(
-      "canvas"
-    )[0];
-    if (canvasRef.current) {
-      view.current = core.createView(canvasRef.current);
-
-      const onResize = () => {
-        if (view.current && canvasRef.current) {
-          canvasRef.current.width =
-            canvasRef.current.clientWidth * window.devicePixelRatio;
-          canvasRef.current.height =
-            canvasRef.current.clientHeight * window.devicePixelRatio;
-          view.current.onResize();
-        }
-      };
-
-      const observer = new ResizeObserver(onResize);
-      observer.observe(canvasRef.current, {});
-
-      const quality =
-        options.detail === "HIGH" ? options.highDetail : options.lowDetail;
-      view.current?.setQuality(...quality);
-
-      setLoaded(true);
-
-      return () => {
-        observer.disconnect();
-        if (canvasContainerRef.current) {
-          canvasContainerRef.current.innerHTML = "";
-          canvasRef.current = undefined;
-        }
-        if (view.current) {
-          view.current.delete();
-          view.current = undefined;
-        }
-      };
-    }
-  }, []);
-
-  React.useEffect(() => {
-    if (view.current) {
-      view.current.showAxes(options.showAxes);
-      view.current.showGrid(options.showGrid);
-      view.current.showWireframe(options.showWireframe);
-      view.current.showShaded(options.showShaded);
-      view.current.setProjection(core.Projection[options.projection]);
-    }
-  }, [
-    loaded,
-    options.showAxes,
-    options.showGrid,
-    options.showShaded,
-    options.showWireframe,
-    options.projection,
-  ]);
-
-  React.useEffect(() => {
-    if (loaded) {
-      const quality =
-        options.detail === "HIGH" ? options.highDetail : options.lowDetail;
-      addTask(
-        () => view.current && core.setRenderQuality(view.current, ...quality)
-      );
-    }
-  }, [options.detail, ...options.highDetail, ...options.lowDetail]);
-
-  return view;
-}
-
-function download(
-  content: string | ArrayBuffer,
-  name: string,
-  mimetype: string
-) {
-  const a = document.createElement("a");
-  a.href = URL.createObjectURL(new Blob([content], { type: mimetype }));
-  a.download = name;
-  a.style.display = "none";
-  document.body.appendChild(a);
-  a.click();
-  URL.revokeObjectURL(a.href);
-  document.body.removeChild(a);
-}
-
-export function useExport(
-  core: ReactCADCore,
-  addTask: AddTask,
-  node: ReactCADNode,
-  name: string | undefined
-): ExportFns {
-  const linearDeflection = 0.05;
-  const angularDeflection = 0.5;
-
-  const exportBREP = React.useCallback(
-    () =>
-      addTask(() => {
-        const progress = core.renderBREP(node);
-
-        progress.then(
-          (content) =>
-            content &&
-            download(content, `${name || "react-cad"}.brep`, "model/brep")
-        );
-
-        return progress;
-      }, false),
-    [core, addTask, node, name]
-  );
-
-  const exportSTEP = React.useCallback(
-    () =>
-      addTask(() => {
-        const progress = core.renderSTEP(node);
-
-        progress.then(
-          (content) =>
-            content &&
-            download(content, `${name || "react-cad"}.step`, "model/step")
-        );
-
-        return progress;
-      }, false),
-    [core, addTask, node, name]
-  );
-
-  const exportSTL = React.useCallback(
-    () =>
-      addTask(() => {
-        const progress = core.renderSTL(
-          node,
-          linearDeflection,
-          angularDeflection
-        );
-
-        progress.then(
-          (content) =>
-            content &&
-            download(content, `${name || "react-cad"}.stl`, "model/stl")
-        );
-
-        return progress;
-      }, false),
-    [core, node, addTask, name]
-  );
-
-  return { exportSTL, exportBREP, exportSTEP };
-}
-
-export function useClickOutside(
-  callback: () => void
-): React.RefObject<HTMLDivElement> {
-  const wrapperRef = React.useRef<HTMLDivElement>(null);
-
-  React.useEffect(() => {
-    function handleClickOutside(event: MouseEvent) {
-      if (
-        wrapperRef.current &&
-        !wrapperRef.current.contains(event.target as Node)
-      ) {
-        callback();
-      }
-    }
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, [wrapperRef]);
-
-  return wrapperRef;
-}
-
-type Task = () => ProgressIndicator<any> | undefined;
-
-type AddTask = (task: Task, cancelPrevious?: boolean) => Promise<void>;
-
-type QueueItem = [Task, () => void];
-
-export function useProgressQueue(): [
-  ProgressIndicator<any> | undefined,
-  AddTask,
-  number
-] {
-  const [progressIndicator, setProgressIndicator] = React.useState<
-    ProgressIndicator<any>
-  >();
-
-  const queue = React.useRef<QueueItem[]>([]);
-
-  const [queueLength, setQueueLength] = React.useState(0);
-
-  const then = React.useCallback(() => {
-    let progress: ProgressIndicator | undefined;
-    let resolve: undefined | (() => void);
-
-    do {
-      const [task, r] = queue.current.shift() ?? [];
-      progress = task?.();
-      if (!progress) {
-        r?.();
-      } else {
-        resolve = r;
-      }
-    } while (queue.current.length && !progress);
-
-    setQueueLength(queue.current.length);
-
-    progress?.then(then, () => {}).then(resolve);
-
-    if (progress) {
-      setProgressIndicator((p) => {
-        p?.cancel();
-        p?.delete();
-        return progress;
-      });
-      progressRef.current = progress;
-    }
-
-    return progress;
-  }, [queue]);
-
-  const addTask = React.useCallback<AddTask>(
-    (task, cancelPrevious = true) => {
-      const queueItem: QueueItem = [task, () => {}];
-      const promise = new Promise<void>((resolve) => (queueItem[1] = resolve));
-      if (
-        !progressIndicator ||
-        progressIndicator.isDeleted() ||
-        progressIndicator.isFulfilled()
-      ) {
-        queue.current.push(queueItem);
-        then();
-      } else {
-        if (cancelPrevious) {
-          queue.current.forEach(([, resolve]) => resolve());
-          queue.current = [queueItem];
-          progressIndicator.catch(then).catch(() => {});
-          progressIndicator.cancel();
-        } else {
-          queue.current.push(queueItem);
-          setQueueLength(queue.current.length);
-        }
-      }
-      return promise;
-    },
-    [progressIndicator, then]
-  );
-
-  // Cancel last task on unmount
-  const progressRef = React.useRef<ProgressIndicator>();
-  React.useEffect(
-    () => () => {
-      if (progressRef.current && !progressRef.current.isDeleted()) {
-        progressRef.current.cancel();
-        progressRef.current.delete();
-      }
-    },
-    []
-  );
-
-  return [progressIndicator, addTask, queueLength];
-}
