+++ packages/viewer/src/hooks/useClickOutside.ts
@@ -1,0 +1,26 @@
+import React from "react";
+
+function useClickOutside(
+  callback: () => void
+): React.RefObject<HTMLDivElement> {
+  const wrapperRef = React.useRef<HTMLDivElement>(null);
+
+  React.useEffect(() => {
+    function handleClickOutside(event: MouseEvent) {
+      if (
+        wrapperRef.current &&
+        !wrapperRef.current.contains(event.target as Node)
+      ) {
+        callback();
+      }
+    }
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, [wrapperRef]);
+
+  return wrapperRef;
+}
+
+export default useClickOutside;
